Setup.py improved (#7)
New features:

Tutorial: This feature is for people who don't know much about Discord bots, or the devoloper mode to be honest. It will tell them how to get certain information on everything that is needed in the setup. Once it's ended it will ask the user if they want to go back to main area(the input that comes when you open up the script).

Setup: This is the old setup that was there before this pull request, nothing has been changed.

Setup & Tutorial(Both): This feature combines the setup with the tutorial, after every question(e.g. Please enter your bot token) there is information on how to get that thing. Such as how to get the token, admin role etc.